Conflicts:
	.gitmodules
	Makefile

+++ extensions/pingback/PingbackController.php
@@ -1,13 +1,25 @@
 <?php
 /**
+ * This file is part of the {@link http://ontowiki.net OntoWiki} project.
+ *
+ * @copyright Copyright (c) 2012, {@link http://aksw.org AKSW}
+ * @license http://opensource.org/licenses/gpl-license.php GNU General Public License (GPL)
+ */
+
+/**
  * semantic pingback controller
  *
  * @category   OntoWiki
- * @package    OntoWiki_extensions_components_pingback
- * @copyright  Copyright (c) 2011, {@link http://aksw.org AKSW}
+ * @package    Extensions_Pingback
+ * @copyright  Copyright (c) 2012, {@link http://aksw.org AKSW}
  * @license    http://opensource.org/licenses/gpl-license.php GNU General Public License (GPL)
+ * @author     Philipp Frischmuth <pfrischmuth@googlemail.com>
+ * @author     Sebastian Tramp <mail@sebastian.tramp.name>
+ * @author     Jonas Brekle <jonas.brekle@gmail.com>
+ * @author     Natanael Arndt <arndtn@gmail.com>
  */
-class PingbackController extends OntoWiki_Controller_Component {
+class PingbackController extends OntoWiki_Controller_Component
+{
 
     protected $_targetGraph = null;
     protected $_sourceRdf = null;
@@ -16,8 +28,12 @@ class PingbackController extends OntoWiki_Controller_Component {
     /**
      * receive a ping
      */
-    public function pingAction() {
+    public function pingAction()
+    {
         $this->_logInfo('Pingback Server Init.');
+
+        $this->_helper->viewRenderer->setNoRender();
+        $this->_helper->layout->disableLayout();
 
         $this->_owApp->appendMessage(
             new OntoWiki_Message('Ping received.', OntoWiki_Message::INFO)
@@ -26,7 +42,7 @@ class PingbackController extends OntoWiki_Controller_Component {
         if (isset($_POST['source']) && isset($_POST['target'])) {
             // Simplified Semantic Pingback
             echo $this->ping($_POST['source'], $_POST['target']);
-            exit;
+            return;
         } else {
             // Create XML RPC Server
             $server = new Zend_XmlRpc_Server();
@@ -35,8 +51,7 @@ class PingbackController extends OntoWiki_Controller_Component {
             // Let the server handle the RPC calls.
             $response = $this->getResponse();
             $response->setBody($server->handle());
-            $response->sendResponse();
-            exit;
+            return;
         }
     }
 
@@ -48,7 +63,8 @@ class PingbackController extends OntoWiki_Controller_Component {
      *
      * @return integer An integer (fault) code
      */
-    public function ping($sourceUri, $targetUri) {
+    public function ping($sourceUri, $targetUri)
+    {
         $this->_logInfo('Method ping was called.');
 
         // Is $targetUri a valid linked data resource in this namespace?
@@ -58,65 +74,47 @@ class PingbackController extends OntoWiki_Controller_Component {
         }
 
         $config = $this->_privateConfig;
-        $foundPingbackTriples = array();
+        $foundPingbackTriplesGraph = array();
 
-        // 1. Try to dereference the source URI as RDF/XML
-        $client = Erfurt_App::getInstance()->getHttpClient($sourceUri, array(
-                    'maxredirects' => 10,
-                    'timeout' => 30
-                ));
-        $client->setHeaders('Accept', 'application/rdf+xml');
-        $client->setHeaders('Content-Type', 'application/rdf+xml');
-        try {
-            $response = $client->request();
-        } catch (Exception $e) {
-            $this->_logError($e->getMessage());
-            return 0x0000;
-        }
-        if ($response->getStatus() === 200) {
-            $data = $response->getBody();
-            $result = $this->_getPingbackTriplesFromRdfXmlString($data, $sourceUri, $targetUri);
-            if (is_array($result)) {
-                $foundPingbackTriples = $result;
-            }
-        }
+        // 1. Try to dereference the source URI as RDF/XML, N3, Truples, Turtle
+        $foundPingbackTriplesGraph = $this->_getResourceFromWrapper($sourceUri, $targetUri, 'Linkeddata');
 
         // 2. If nothing was found, try to use as RDFa service
-        if (((boolean) $config->rdfa->enabled) && (count($foundPingbackTriples) === 0)) {
-            $service = $config->rdfa->service . urlencode($sourceUri);
-            $client = Erfurt_App::getInstance()->getHttpClient($service, array(
-                        'maxredirects' => 10,
-                        'timeout' => 30
-                    ));
+        if (((boolean) $config->rdfa->enabled) && (count($foundPingbackTriplesGraph) === 0)) {
+            $foundPingbackTriplesGraph = $this->_getResourceFromWrapper($sourceUri, $targetUri, 'Rdfa');
+        }
 
-            try {
-                $response = $client->request();
-            } catch (Exception $e) {
-                $this->_logError($e->getMessage());
-                return 0x0000;
-            }
-            if ($response->getStatus() === 200) {
-                $data = $response->getBody();
-                $result = $this->_getPingbackTriplesFromRdfXmlString($data, $sourceUri, $targetUri);
-                if ($result) {
-                    $foundPingbackTriples = $result;
+        $foundPingbackTriples = array();
+        foreach ($foundPingbackTriplesGraph as $s => $predicates) {
+            foreach ($predicates as $p => $objects) {
+                foreach ($objects as $o) {
+                    $foundPingbackTriples[] = array(
+                        's' => $s,
+                        'p' => $p,
+                        'o' => $o['value']
+                    );
                 }
             }
         }
 
         $versioning = Erfurt_App::getInstance()->getVersioning();
-        $versioning->startAction(array(
-            'type' => '9000',
-            'modeluri' => $this->_targetGraph,
-            'resourceuri' => $sourceUri
-        ));
+        $versioning->startAction(
+            array(
+                'type' => '9000',
+                'modeluri' => $this->_targetGraph,
+                'resourceuri' => $sourceUri
+            )
+        );
 
         // 3. If still nothing was found, try to find a link in the html
         if (count($foundPingbackTriples) === 0) {
-            $client = Erfurt_App::getInstance()->getHttpClient($sourceUri, array(
-                        'maxredirects' => 10,
-                        'timeout' => 30
-                    ));
+            $client = Erfurt_App::getInstance()->getHttpClient(
+                $sourceUri,
+                array(
+                    'maxredirects' => 10,
+                    'timeout' => 30
+                )
+            );
 
             try {
                 $response = $client->request();
@@ -188,18 +186,20 @@ class PingbackController extends OntoWiki_Controller_Component {
         return 'Pingback has been registered or updated... Keep spinning the Data Web ;-)';
     }
 
-    protected function _addPingback($s, $p, $o) {
+    protected function _addPingback($s, $p, $o)
+    {
         if ($this->_targetGraph === null) {
             return false;
         }
 
         $store = Erfurt_App::getInstance()->getStore();
 
-        $sql = 'INSERT INTO ow_pingback_pingbacks (source, target, relation) VALUES ("' . $s . '", "' . $o . '", "' . $p . '")';
+        $sql = 'INSERT INTO ow_pingback_pingbacks (source, target, relation) '
+            . 'VALUES ("' . $s . '", "' . $o . '", "' . $p . '")';
         $this->_query($sql);
 
         $store->addStatement(
-                $this->_targetGraph, $s, $p, array('value' => $o, 'type' => 'uri'), false
+            $this->_targetGraph, $s, $p, array('value' => $o, 'type' => 'uri'), false
         );
 
         if ($this->_sourceRdf !== null) {
@@ -207,7 +207,7 @@ class PingbackController extends OntoWiki_Controller_Component {
                 $titleProps = $this->_privateConfig->title_properties->toArray();
                 if (in_array($prop, $titleProps)) {
                     $store->addStatement(
-                            $this->_targetGraph, $s, $prop, $oArray[0], false
+                        $this->_targetGraph, $s, $prop, $oArray[0], false
                     );
                     break; // only one title
                 }
@@ -220,11 +220,11 @@ class PingbackController extends OntoWiki_Controller_Component {
         $event->o = $o;
         $event->trigger();
 
-
         return true;
     }
 
-    protected function _checkTargetExists($targetUri) {
+    protected function _checkTargetExists($targetUri)
+    {
         if ($this->_targetGraph == null) {
             $event = new Erfurt_Event('onNeedsGraphForLinkedDataUri');
             $event->uri = $targetUri;
@@ -241,44 +241,57 @@ class PingbackController extends OntoWiki_Controller_Component {
         }
     }
 
-    function _deleteInvalidPingbacks($sourceUri, $targetUri, $foundPingbackTriples = array()) {
+    function _deleteInvalidPingbacks($sourceUri, $targetUri, $foundPingbackTriples = array())
+    {
         $store = Erfurt_App::getInstance()->getStore();
 
         $sql = 'SELECT * FROM ow_pingback_pingbacks WHERE source="' . $sourceUri . '" AND target="' . $targetUri . '"';
         $result = $this->_query($sql);
 
         $removed = false;
-        foreach ($result as $row) {
-            $found = false;
-            foreach ($foundPingbackTriples as $triple) {
-                if ($triple['p'] === $row['relation']) {
-                    $found = true;
-                    break;
+        if ($result !== false) {
+            foreach ($result as $row) {
+                $found = false;
+                foreach ($foundPingbackTriples as $triple) {
+                    if ($triple['p'] === $row['relation']) {
+                        $found = true;
+                        break;
+                    }
                 }
-            }
 
-            if (!$found) {
-                $sql = 'DELETE FROM ow_pingback_pingbacks WHERE id=' . $row['id'];
-                $this->_query($sql);
+                if (!$found) {
+                    $sql = 'DELETE FROM ow_pingback_pingbacks WHERE id=' . $row['id'];
+                    $this->_query($sql);
 
-                $oSpec = array(
-                    'value' => $targetUri,
-                    'type' => 'uri'
-                );
+                    $oSpec = array(
+                        'value' => $targetUri,
+                        'type' => 'uri'
+                    );
 
-                $store->deleteMatchingStatements($this->_targetGraph, $sourceUri, $row['relation'], $oSpec, array('use_ac' => false));
-                $removed = true;
+                    $store->deleteMatchingStatements(
+                        $this->_targetGraph,
+                        $sourceUri,
+                        $row['relation'],
+                        $oSpec,
+                        array('use_ac' => false)
+                    );
+                    $removed = true;
+                }
             }
         }
 
         return $removed;
     }
 
-    protected function _determineInverseProperty($propertyUri) {
-        $client = Erfurt_App::getInstance()->getHttpClient($propertyUri, array(
-                    'maxredirects' => 10,
-                    'timeout' => 30
-                ));
+    protected function _determineInverseProperty($propertyUri)
+    {
+        $client = Erfurt_App::getInstance()->getHttpClient(
+            $propertyUri,
+            array(
+                'maxredirects' => 10,
+                'timeout' => 30
+            )
+        );
         $client->setHeaders('Accept', 'application/rdf+xml');
         try {
             $response = $client->request();
@@ -307,73 +320,63 @@ class PingbackController extends OntoWiki_Controller_Component {
         }
     }
 
-    protected function _getPingbackTriplesFromRdfXmlString($rdfXml, $sourceUri, $targetUri) {
-        $parser = Erfurt_Syntax_RdfParser::rdfParserWithFormat('rdfxml');
-        try {
-            $result = $parser->parse($rdfXml, Erfurt_Syntax_RdfParser::LOCATOR_DATASTRING);
-        } catch (Exception $e) {
-            $this->_logError($e->getMessage());
-            return false;
+    private function _getResourceFromWrapper($sourceUri, $targetUri, $wrapperName = 'Linkeddata')
+    {
+        $r = new Erfurt_Rdf_Resource($sourceUri);
+
+        // Try to instanciate the requested wrapper
+        new Erfurt_Wrapper_Manager();
+        $wrapper = Erfurt_Wrapper_Registry::getInstance()->getWrapperInstance($wrapperName);
+
+        $wrapperResult = null;
+        $wrapperResult = $wrapper->run($r, null, true);
+
+        $newStatements = null;
+        if ($wrapperResult === false) {
+            // IMPORT_WRAPPER_NOT_AVAILABLE;
+        } else if (is_array($wrapperResult)) {
+            $newStatements = $wrapperResult['add'];
+            // TODO make sure to only import the specified resource
+            $newModel = new Erfurt_Rdf_MemoryModel($newStatements);
+            $newStatements = array();
+            $object = array('type' => 'uri', 'value' => $targetUri);
+            $newStatements = $newModel->getP($sourceUri, $object);
+        } else {
+            // IMPORT_WRAPPER_ERR;
         }
 
-        if (isset($result[$sourceUri])) {
-            $this->_sourceRdf = $result[$sourceUri];
-        }
-
-        $foundTriples = array();
-        foreach ($result as $s => $pArray) {
-            foreach ($pArray as $p => $oArray) {
-                foreach ($oArray as $oSpec) {
-                    if ($s === $sourceUri) {
-                        if (($oSpec['type'] === 'uri') && ($oSpec['value'] === $targetUri)) {
-                            $foundTriples[] = array(
-                                's' => $s,
-                                'p' => $p,
-                                'o' => $oSpec['value']
-                            );
-                        }
-                    } else if (($oSpec['type'] === 'uri') && ($oSpec['value'] === $sourceUri)) {
-                        // Try to find inverse property for $p
-                        $inverseProp = $this->_determineInverseProperty($p);
-                        if ($inverseProp !== null) {
-                            $foundTriples[] = array(
-                                's' => $oSpec['value'],
-                                'p' => $inverseProp,
-                                'o' => $s
-                            );
-                        }
-                    }
-                }
-            }
-        }
-
-        return $foundTriples;
+        return $newStatements;
     }
 
-    protected function _logError($msg) {
+    protected function _logError($msg)
+    {
         $owApp = OntoWiki::getInstance();
         $logger = $owApp->logger;
 
         if (is_array($msg)) {
-            $logger->debug('Pingback Component Error: ' . print_r($msg, true));
+            $logger->debug('Pingback Component Error: ' . var_export($msg, true));
         } else {
             $logger->debug('Pingback Component Error: ' . $msg);
         }
     }
 
-    protected function _logInfo($msg) {
+    protected function _logInfo($msg)
+    {
         $owApp = OntoWiki::getInstance();
         $logger = $owApp->logger;
 
         if (is_array($msg)) {
-            $logger->debug('Pingback Component Info: ' . print_r($msg, true));
+            $logger->debug('Pingback Component Info: ' . var_export($msg, true));
         } else {
             $logger->debug('Pingback Component Info: ' . $msg);
         }
     }
 
-    protected function _pingbackExists($s, $p, $o) {
-        $sql = 'SELECT * FROM ow_pingback_pingbacks WHERE source="' . $s . '" AND target="' . $o . '" AND relation="' . $p . '" LIMIT 1';
+    protected function _pingbackExists($s, $p, $o)
+    {
+        $sql = 'SELECT * FROM ow_pingback_pingbacks '
+            . 'WHERE source="' . $s . '" AND target="' . $o . '" AND relation="' . $p . '" '
+            . 'LIMIT 1';
         $result = $this->_query($sql);
         if (is_array($result) && (count($result) === 1)) {
             return true;
@@ -382,7 +385,8 @@ class PingbackController extends OntoWiki_Controller_Component {
         return false;
     }
 
-    private function _checkDb() {
+    private function _checkDb()
+    {
         if ($this->_dbChecked) {
             return;
         }
@@ -399,7 +403,8 @@ class PingbackController extends OntoWiki_Controller_Component {
         $this->_dbChecked = true;
     }
 
-    private function _createTable() {
+    private function _createTable()
+    {
         $store = Erfurt_App::getInstance()->getStore();
 
         $sql = 'CREATE TABLE IF NOT EXISTS ow_pingback_pingbacks (
@@ -412,7 +417,8 @@ class PingbackController extends OntoWiki_Controller_Component {
         return $this->_query($sql, false);
     }
 
-    protected function _query($sql, $withCheck = true) {
+    protected function _query($sql, $withCheck = true)
+    {
         if ($withCheck) {
             $this->_checkDb();
         }
@@ -427,5 +433,4 @@ class PingbackController extends OntoWiki_Controller_Component {
 
         return $result;
     }
-
 }
